Cleanup TV related code in API
