+++ src/NzbDrone.Api/History/HistoryModule.cs
@@ -1,7 +1,7 @@
-﻿using System;
+using System;
 using Nancy;
 using NzbDrone.Api.Extensions;
-using NzbDrone.Api.Movie;
+using NzbDrone.Api.Movies;
 using NzbDrone.Core.Datastore;
 using NzbDrone.Core.DecisionEngine;
 using NzbDrone.Core.Download;
