+++ src/NzbDrone.Api/Movies/AlternativeYearModule.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using Marr.Data;
 using Nancy;
 using NzbDrone.Api;
-using NzbDrone.Api.Movie;
 using NzbDrone.Common.Cache;
 using NzbDrone.Common.Extensions;
 using NzbDrone.Common.Messaging;
@@ -19,7 +18,7 @@ using NzbDrone.Core.RootFolders;
 using NzbDrone.Core.Tv;
 using NzbDrone.Core.Tv.Events;
 
-namespace NzbDrone.Api.Movie
+namespace NzbDrone.Api.Movies
 {
     public class AlternativeYearModule : NzbDroneRestModule<AlternativeYearResource>
     {
