+++ src/NzbDrone.Api/Movies/MovieEditorModule.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Nancy;
@@ -7,7 +7,7 @@ using NzbDrone.Api.Extensions;
 using NzbDrone.Api.REST;
 using NzbDrone.Core.Tv;
 
-namespace NzbDrone.Api.Movie
+namespace NzbDrone.Api.Movies
 {
     public class MovieEditorModule : NzbDroneApiModule
     {
