+++ src/NzbDrone.Api/NetImport/ImportExclusionsResource.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Linq;
 using NzbDrone.Core.NetImport;
 using NzbDrone.Core.Tv;
