+++ src/NzbDrone.Api/NzbDrone.Api.csproj
@@ -126,8 +126,8 @@
     <Compile Include="Movies\AlternativeYearModule.cs" />
     <Compile Include="Movies\MovieModuleWithSignalR.cs" />
     <Compile Include="Movies\MovieBulkImportModule.cs" />
-    <Compile Include="Movies\MovieFileModule.cs" />
-    <Compile Include="Series\MovieModule.cs" />
+    <Compile Include="MovieFiles\MovieFileModule.cs" />
+    <Compile Include="Movies\MovieModule.cs" />
     <Compile Include="Movies\RenameMovieModule.cs" />
     <Compile Include="Movies\RenameMovieResource.cs" />
     <Compile Include="Movies\MovieEditorModule.cs" />
@@ -162,13 +162,6 @@
     <Compile Include="DiskSpace\DiskSpaceResource.cs" />
     <Compile Include="DownloadClient\DownloadClientModule.cs" />
     <Compile Include="DownloadClient\DownloadClientResource.cs" />
-    <Compile Include="EpisodeFiles\EpisodeFileModule.cs" />
-    <Compile Include="EpisodeFiles\EpisodeFileResource.cs" />
-    <Compile Include="Episodes\EpisodeModule.cs" />
-    <Compile Include="Episodes\EpisodeModuleWithSignalR.cs" />
-    <Compile Include="Episodes\EpisodeResource.cs" />
-    <Compile Include="Episodes\RenameEpisodeModule.cs" />
-    <Compile Include="Episodes\RenameEpisodeResource.cs" />
     <Compile Include="ErrorManagement\ApiException.cs" />
     <Compile Include="ErrorManagement\ErrorHandler.cs" />
     <Compile Include="ErrorManagement\ErrorModel.cs" />
@@ -244,17 +237,15 @@
     <Compile Include="REST\RestResource.cs" />
     <Compile Include="RootFolders\RootFolderModule.cs" />
     <Compile Include="RootFolders\RootFolderResource.cs" />
-    <Compile Include="SeasonPass\SeasonPassResource.cs" />
-    <Compile Include="Series\AlternativeTitleResource.cs" />
-    <Compile Include="Series\MovieFileResource.cs" />
-    <Compile Include="Series\FetchMovieListModule.cs" />
+    <Compile Include="Movies\AlternativeTitleResource.cs" />
+    <Compile Include="MovieFiles\MovieFileResource.cs" />
+    <Compile Include="Movies\FetchMovieListModule.cs" />
     <Compile Include="Series\SeasonResource.cs" />
-    <Compile Include="SeasonPass\SeasonPassModule.cs" />
     <Compile Include="Series\SeriesEditorModule.cs" />
-    <Compile Include="Series\MovieLookupModule.cs" />
+    <Compile Include="Movies\MovieLookupModule.cs" />
     <Compile Include="Series\SeriesLookupModule.cs" />
     <Compile Include="Series\SeriesModule.cs" />
-    <Compile Include="Series\MovieResource.cs" />
+    <Compile Include="Movies\MovieResource.cs" />
     <Compile Include="Series\SeriesResource.cs" />
     <Compile Include="Series\SeasonStatisticsResource.cs" />
     <Compile Include="System\Backup\BackupModule.cs" />
@@ -271,12 +262,10 @@
     <Compile Include="Validation\RssSyncIntervalValidator.cs" />
     <Compile Include="Validation\EmptyCollectionValidator.cs" />
     <Compile Include="Validation\RuleBuilderExtensions.cs" />
-    <Compile Include="Wanted\CutoffModule.cs" />
     <Compile Include="Wanted\LegacyMissingModule.cs" />
-    <Compile Include="Wanted\MissingModule.cs" />
     <Compile Include="Wanted\MovieCutoffModule.cs" />
     <Compile Include="Wanted\MovieMissingModule.cs" />
-    <Compile Include="Series\MovieDiscoverModule.cs" />
+    <Compile Include="Movies\MovieDiscoverModule.cs" />
     <Compile Include="NetImport\ImportExclusionsModule.cs" />
     <Compile Include="NetImport\ImportExclusionsResource.cs" />
   </ItemGroup>
