+++ src/NzbDrone.Api/Parse/ParseResource.cs
@@ -1,7 +1,6 @@
-﻿using System.Collections.Generic;
-using NzbDrone.Api.Episodes;
+using System.Collections.Generic;
+using NzbDrone.Api.Movies;
 using NzbDrone.Api.REST;
-using NzbDrone.Api.Series;
 using NzbDrone.Core.Parser.Model;
 
 namespace NzbDrone.Api.Parse
@@ -9,8 +8,7 @@ namespace NzbDrone.Api.Parse
     public class ParseResource : RestResource
     {
         public string Title { get; set; }
-        public ParsedEpisodeInfo ParsedEpisodeInfo { get; set; }
-        public SeriesResource Series { get; set; }
-        public List<EpisodeResource> Episodes { get; set; }
+        public ParsedMovieInfo ParsedMovieInfo { get; set; }
+        public MovieResource Movie { get; set; }
     }
 }
