+++ src/NzbDrone.Api/Queue/QueueResource.cs
@@ -1,10 +1,8 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using NzbDrone.Api.REST;
 using NzbDrone.Core.Qualities;
-using NzbDrone.Api.Series;
-using NzbDrone.Api.Episodes;
-using NzbDrone.Api.Movie;
+using NzbDrone.Api.Movies;
 using NzbDrone.Core.Download.TrackedDownloads;
 using NzbDrone.Core.Indexers;
 using System.Linq;
@@ -13,8 +11,6 @@ namespace NzbDrone.Api.Queue
 {
     public class QueueResource : RestResource
     {
-        public SeriesResource Series { get; set; }
-        public EpisodeResource Episode { get; set; }
         public MovieResource Movie { get; set; }
         public QualityModel Quality { get; set; }
         public decimal Size { get; set; }
@@ -38,9 +34,6 @@ namespace NzbDrone.Api.Queue
             return new QueueResource
             {
                 Id = model.Id,
-
-                Series = model.Series.ToResource(),
-                Episode = model.Episode.ToResource(),
                 Quality = model.Quality,
                 Size = model.Size,
                 Title = model.Title,
