+++ src/NzbDrone.Api/Wanted/MovieMissingModule.cs
@@ -1,4 +1,3 @@
-﻿using NzbDrone.Api.Movie;
 using NzbDrone.Api.Movies;
 using NzbDrone.Core.DecisionEngine;
 using NzbDrone.Core.Tv;
