+++ src/NzbDrone.Integration.Test/ApiTests/BlacklistFixture.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using NUnit.Framework;
-using NzbDrone.Api.Movie;
+using NzbDrone.Api.Movies;
 
 namespace NzbDrone.Integration.Test.ApiTests
 {
